Run the tag browser tests on every desktop project
The clipboard tests were scoped to Chrome because granting the clipboard
permissions throws `Unknown permission: clipboard-read` on Firefox. Only
the grant is Chromium specific though: the tests write with
`navigator.clipboard.writeText`, which a focused document is allowed to
do, and paste with a real key press. Guarding the grant lets the whole
file run on Chrome, Firefox and Safari, so the tag list structure test
moves in here rather than living in a second file.
